refactor(ui/web): migrate PositionsList, OrderSummary, and PositionsTable rows from @starboard
