--- frontend/src/@starboard/pages/Dashboard/submodules/PositionsTable/components/Row/components/index.ts
@@ -1,5 +1,0 @@
-export { FeesCell } from './FeesCell';
-export { PnlDeltaCell } from './PnlDeltaCell';
-export { SizeDeltaCell } from './SizeDeltaCell';
-export { TimestampCell } from './TimestampCell';
-export { TradeValueCell } from './TradeValueCell';
