+++ frontend/src/pages/Dashboard/src/submodules/IncreasePositionForm/src/components/index.ts
@@ -5,4 +5,4 @@ export * from './LeverageInput';
 export * from './OrderSideSwitch';
 export * from './PositionSizeInputs';
 export * from './SubmitButton';
-export * from '@/@starboard/pages/Dashboard/submodules/IncreasePositionForm/components/OrderSummary';
+export * from './OrderSummary';
