+++ frontend/src/pages/Dashboard/src/submodules/PositionsTable/components/Row/components/PnlDeltaCell.tsx
@@ -1,9 +1,9 @@
 import { type FC, use } from 'react';
 import { $decimalValue } from 'fuel-ts-sdk';
 import { formatCurrency } from '@/lib/formatCurrency';
-import * as $ from '@/pages/Dashboard/src/submodules/PositionsTable/components/Row/components/RealizedPnlCell.css';
-import { Cell } from '@/pages/Dashboard/src/submodules/PositionsTable/components/Row/components/common/Cell';
-import { RowContext } from '@/pages/Dashboard/src/submodules/PositionsTable/components/Row/contexts/RowContext';
+import { RowContext } from '../contexts/RowContext';
+import * as $ from './RealizedPnlCell.css';
+import { Cell } from './common/Cell';
 
 export const PnlDeltaCell: FC = () => {
   const position = use(RowContext)!;
