+++ frontend/src/pages/Dashboard/src/submodules/PositionsTable/components/Row/components/TimestampCell.tsx
@@ -1,6 +1,6 @@
 import { type FC, use } from 'react';
-import { Cell } from '@/pages/Dashboard/src/submodules/PositionsTable/components/Row/components/common/Cell';
-import { RowContext } from '@/pages/Dashboard/src/submodules/PositionsTable/components/Row/contexts/RowContext';
+import { RowContext } from '../contexts/RowContext';
+import { Cell } from './common/Cell';
 
 export const TimestampCell: FC = () => {
   const position = use(RowContext)!;
