+++ frontend/src/pages/Dashboard/src/submodules/PositionsTable/components/Row/components/index.ts
@@ -1,15 +1,12 @@
 export { ActionCell } from './ActionCell';
 export { CollateralCell } from './CollateralCell';
-export type * from './common/Cell';
+export { FeesCell } from './FeesCell';
 export { NetValueCell } from './NetValueCell';
+export { PnlDeltaCell } from './PnlDeltaCell';
 export { PositionCell } from './PositionCell';
 export { PriceCell } from './PriceCell';
 export { RealizedPnlCell } from './RealizedPnlCell';
 export { SizeCell } from './SizeCell';
-
-// New cells from @starboard/
-export { FeesCell } from '@/@starboard/pages/Dashboard/submodules/PositionsTable/components/Row/components/FeesCell';
-export { PnlDeltaCell } from '@/@starboard/pages/Dashboard/submodules/PositionsTable/components/Row/components/PnlDeltaCell';
-export { SizeDeltaCell } from '@/@starboard/pages/Dashboard/submodules/PositionsTable/components/Row/components/SizeDeltaCell';
-export { TimestampCell } from '@/@starboard/pages/Dashboard/submodules/PositionsTable/components/Row/components/TimestampCell';
-export { TradeValueCell } from '@/@starboard/pages/Dashboard/submodules/PositionsTable/components/Row/components/TradeValueCell';
+export { SizeDeltaCell } from './SizeDeltaCell';
+export { TimestampCell } from './TimestampCell';
+export { TradeValueCell } from './TradeValueCell';
